dwarfx: handle multiple EndSequence entries per line table
Upon more careful reading of the DWARF standard and the thread at
http://lists.celinuxforum.org/pipermail/dwarf-workgroup/2005-August/000216.html
it's clear that EndSequence entries can appear in the middle of a line
table, in addition to the required entry at the end of a line table.

Fix the line table reader to account for this.  The previous
assumption was woven through many parts of the implementation and
interface, so this proved somewhat tricky.  This changes the interface
to Next, since it needs a way to indicate a clean end-of-table other
than EndSequence.  Next also needs a more complete implementation of
lneEndSequence, since more rows can follow.  We also need to handle
the end of the buffer more carefully, since that now indicates the end
of the line table.  Finally, the interface to SeekPC also changes,
since a PC can fall in a hole in the middle of a line table without
coming either before or after the table as a whole.