cmd/memlat: always show all tabs
Currently the browser only shows valid tabs and changes both the
visible tabs and the selected tabs as filters are applied. This is
very confusing because it's not obvious the tabs are changing or what
tabs can be made available.

Instead, always show all of the tabs, but disable selecting rows on
tabs that already have a filter and show an explanation on the
annotation tab when no function has been selected.