cmd/memlat: compute labels in filter bar on demand
Currently the memlat table computes labels for the rows and then
communicates these along with the filter to the filter bar via a
property. This has the annoying side-effect of putting these redundant
labels in the hash state.

Improve this by communicating only the filter via properties and the
hash state and simply computing the filter label via a function shared
between the memlat table and the filter bar.