perffile: support ABI v4 perf_event_attr and fix regs parsing
ABI v4, introduced in Linux 3.18, adds support for capturing
interrupt-time registers in samples in addition to user-space
registers. This commit adds support for parsing this in the event
structure and the sample structure.

This also fixes the parsing of both user-space and interrupt-time
registers when a sample specifies an ABI of "none". This indicates
that there are no registers in this sample, even if the event
specifies registers, but previously the sample parser was attempting
to read registers regardless of the sample ABI.