perffile: split unioned fields in EventAttr
Previously EventAttr needed to correspond to the on-disk
perf_event_attr structure. Now there's a separate, private structure
for that, so EventAttr can be more user-friendly.

Several of the fields in EventAttr represent unions in the C
structure. Split these out and figure out which field is appropriate
when converting the on-disk representation in to the exported
representation.