perfsession: use ELF symbols if there are zero DWARF symbols
When DWARF info is split in to a DWARF package file (.dwp), the
original binary retains a .debug_info section that references the .dwp
file but doesn't actually contain function information. Currently this
confuses symbolization because it find the .debug_info section,
assumes it can get function symbols from it, and ignores the ELF
symbol table entirely.

Fix this by falling back to the ELF symbol table if there's either no
.debug_info section *or* the .debug_info section contains no function
symbols.