perffile: accept missing eventAttr on some events
perf recently started omitting the eventAttr for at least mmap events
in the file prologue. This isn't too surprising, since the PID and TID
are the only fields that make sense in these and they're also in the
main mmap event, but it broke the parser. Fix this by accepting
missing eventAttrs on mmap events.