memlat: new source annotation to handle inlining
The current source annotation implementation assumes all samples will
be within a single function. However, if there's inlining, it will
pick the source file with the highest-weight histogram and assume all
line numbers are for that file.

Fix this by generalizing the implementation to simply show the regions
of source surrounding each sampled line.