perffile: add event PMU type id to hardware events
Added upstream in 55bcf6ef314ae8ba81bcd74aa760247b635ed47b ("perf:
Extend PERF_TYPE_HARDWARE and PERF_TYPE_HW_CACHE").

This isn't conditional on the PMU capabilities, as those are not exposed
in the perf file as far as I can tell.

The PMUTypeID is mapped back to its name with FileMeta.PMUMappings,
which should not be type EventType.

Evidence:

* PMU type ID is set to pmu->type:
  https://elixir.bootlin.com/linux/v5.18.9/source/tools/perf/util/parse-events-hybrid.c#L51

* pmu->type is set with the contents of
  /sys/bus/event_source/devices/$PMU/type:
  https://elixir.bootlin.com/linux/v5.18.9/source/tools/perf/util/pmu.c#L562

* pmu->type is the value in the PMU mappings:
  https://elixir.bootlin.com/linux/v5.18.9/source/tools/perf/util/header.c#L774